Fix typo in upper constraints
This fixes running tox tests locally.

Note, by default this still uses the most recent magnum tag
which is not the same as what happens in the gate right now.
There might be future issues here as upper constraints
changes in ways that require fixes in the core of Magnum.

Change-Id: I76db3b20e29ce2f445e871ee4e8083855c85fa37