Support autoscaling for the default node group
Similar to the heat driver, this patch means we now respect the cluster labels that specifiy a min and max node count for the default worker group.

Note that (unlike the heat driver) the "current" node count is being ignored here, and the cluster always starts with min_node_count in the default group, if that label has been specified.

Change-Id: Ic0e22efdf2ccb06b19ea713413eb3b6cfc405de4