Use API LB floating IPs according to template and label
Label master_lb_floating_ip_enabled is used if present.
Cluster template field floating_ip_enabled is used
as a fallback field because it has a wider scope (it also
means to add a floating ip to all worker nodes - not implemented
in the Magnum CAPI Helm driver).

When this option is disabled the CAPI Management cluster
must have an alternative route to reach the KubeAPI,
otherwise cluster creation will not succeed.

Closes-bug: #2085339
Co-Authored-By: Piotr Parczewski <piotr@stackhpc.com>
Change-Id: I48fda8cdb6e9de915528e47de40a771614c092d1
Signed-off-by: Dale Smith <dale@catalystcloud.nz>