Start using csi_cinder_volume_binding_mode config
By default, since release 0.8.0 of capi-helm, it has
been defaulting to WaitForFirstConsumer for additional
storage classes.
https://github.com/azimuth-cloud/capi-helm-charts/commit/854172966653138a60c329215fe9bb18726353e6

Note, if you are upgrading from an older version,
we hit some upgrade issues due to immutable storage
classes being unable to be updated.

As a work around, we are actully making use of the
existing config, to keep the behaviour the same as
the current release of the capi-helm-charts.
But those using older versions, before 0.8.0, may
need to change the configuration to match their existing
choices, ahead of their upgrade.

Change-Id: If90e0143f56ee378f5401b01623c159836fef29c
Signed-off-by: John Garbutt <john.garbutt@stackhpc.com>