Add locking mechanism to Helm install/upgrade commands
Multiple concurrent Helm operations, such as creating many nodegroups
at the same time, could lead to conflicts when updating the Helm
release. Resolve this with a locking mechanism which uses the
Kubernetes lease objects for co-ordination.

This is an alternative to the Helm reconciliation loop approach
proposed in [1] which is simpler to reason about and has the advantage
that Helm install/upgrade commands are eagerly executed rather than
delayed until the next reconciliation loop iteration.

One known drawback of this approach is that if a Magnum conductor
process is terminated while waiting for a lock, the pending update
to a cluster's state will be persisted to the DB but the Helm values
will not be applied until another Helm operation is executed against
the same cluster.
This could be resolved in a follow up patch which applies all Helm
updates to existing clusters during driver initialisation.

[1] https://review.opendev.org/c/openstack/magnum-capi-helm/+/946055

Closes-bug: #2097946
Change-Id: I46ad5088072f379c55b21646455e6d4b87da2f15
Signed-off-by: Scott Davidson <scott@stackhpc.com>
Signed-off-by: Michal Nasiadka <mnasiadka@gmail.com>